New test workflows and support
There's a new poetry image with out dynamic versioning and with
poetry installed in a venv so that we can install dynamic versioning
in to that venv when needed.
There's composite action at OxIonics/poetry-preamble for us to put
poetry setup stuff in.
There's test.yml workflow for running just the tests and a
test_flake_yapf.yml workflow for running all 3 of them.