Fix image clean-up (#52)
We want to tag the image at the beginning and remove that tag at the end. If the image was not tagged with another tag (which we do if triggered by a push), this removes the image itself, too.
But we want a reliably unique tag as the temporary tag name. Using the build id alone is not necessarily unique, because some workflows call this workflow multiple times.
The image id generated by Docker is unique - so let's use that to construct a temporary tag name.