Don't log into container registry from multiple jobs in parallel
The `docker/login-action` automatically adds a log out step at the end of the job.
Sadly, this seems to also affect any other job that runs in parallel and has logged into the registry.
We only need to log in when we intend to push an image to the registry, so in those cases we can set a concurrency group to avoid running this job multiple times in parallel.