use `poe lint-rs` to run clippy
Was having issues with giraffate/clippy-action failing silently that i
couldn't quickly get to the bottom of. In general i think it would be
better for us to use less third party actions to avoid such issues.
Plus, this way we don't duplicate the clippy flags between
`pyproject.toml` and here.

This does assume that rust projects using this will have a
`pyproject.toml` with poe tasks defined, which is currently true, and
likely to always be true as we use it for the base project
configuration.

This requires https://github.com/OxIonics/base_project/pull/195